define new haplotype class for haptools
